requirements.txt
asgiref==3.8.1
Django==5.1
django-unfold==0.34.0
sqlparse==0.5.1
